Use a serial queue to prevent open/close race conditions

+++ Mongod.js
@@ -17,6 +17,7 @@
 
 const childprocess = require('child_process');
 const events = require('events');
+const PromiseQueue = require('promise-queue');
 const keyRE = /(pid=\d+)|(port=\d+)|(waiting\s+for\s+connections)|(already\s+in\s+use)|(denied\s+for\s+socket)|((error|exception|badvalue)(.|\n)*)/ig;
 const errorRE = /^error|exception|badvalue/i;
 const whiteSpaceRE = / /ig;
@@ -122,138 +123,131 @@ class Mongod extends events.EventEmitter {
    * Start a given {@link Mongod}.
    * @protected
    * @argument {Mongod} server
-   * @argument {Boolean} isCallback
    * @return {Promise}
    */
-  static open(server, isCallback) {
-    if (server.isClosing) {
-      server.openPromise = new Promise((resolve, reject) => {
-        const open = () =>
-          Mongod.open(server, true).then(resolve).catch(reject);
-
-        server.isClosing = false;
-        server.isOpening = true;
-
-        server.closePromise.then(open).catch(open);
-      });
-
+  static open(server) {
+    if (server.isOpening) {
       return server.openPromise;
     }
 
-    if (server.isOpening && !isCallback || server.isRunning) {
-      return server.openPromise;
-    }
-
-    server.openPromise = new Promise((resolve, reject) => {
-      const matchHandler = (match) => {
-        let err = null;
-        let k, v;
-
-        if (errorRE.test(match)) {
-          k = 'error';
-          v = match.trim();
-        }
-        else {
-          const t = match.split('=');
-
-          k = t[0].replace(whiteSpaceRE, '').toLowerCase();
-          v = t[1];
-        }
-
-        switch (k) {
-          case 'error':
-            err = new Error(v);
-            err.code = -1;
-
-            break;
-
-          case 'alreadyinuse':
-            err = new Error('Address already in use');
-            err.code = -2;
-
-            break;
-
-          case 'deniedforsocket':
-            err = new Error('Permission denied');
-            err.code = -3;
-
-            break;
-
-          case 'pid':
-          case 'port':
-            server[k] = Number(v);
-
-            return false;
-
-          case 'waitingforconnections':
-            server.isRunning = true;
-
-            server.emit('open');
-
-            break;
-
-          default:
-            return false;
-        }
-
+    server.isOpening = true;
+    server.isClosing = false;
+    server.openPromise = server.promiseQueue.add(() => {
+      if (server.isClosing || server.isRunning) {
         server.isOpening = false;
 
-        if (err === null) {
-          resolve(null);
-        }
-        else {
-          reject(err);
-        }
+        return Promise.resolve(null);
+      }
 
-        return true;
-      };
+      return new Promise((resolve, reject) => {
+        const matchHandler = (match) => {
+          let err = null;
+          let k, v;
 
-      const dataHandler = Mongod.getTextLineAggregator((value) => {
-        const matches = value.match(keyRE);
+          if (errorRE.test(match)) {
+            k = 'error';
+            v = match.trim();
+          }
+          else {
+            const t = match.split('=');
 
-        if (matches !== null) {
-          for (let match of matches) {
-            if (matchHandler(match, value)) {
-              server.process.stdout.removeListener('data', dataHandler);
-              server.process.stderr.removeListener('data', dataHandler);
+            k = t[0].replace(whiteSpaceRE, '').toLowerCase();
+            v = t[1];
+          }
 
-              return;
+          switch (k) {
+            case 'error':
+              err = new Error(v);
+              err.code = -1;
+
+              break;
+
+            case 'alreadyinuse':
+              err = new Error('Address already in use');
+              err.code = -2;
+
+              break;
+
+            case 'deniedforsocket':
+              err = new Error('Permission denied');
+              err.code = -3;
+
+              break;
+
+            case 'pid':
+            case 'port':
+              server[k] = Number(v);
+
+              return false;
+
+            case 'waitingforconnections':
+              server.isRunning = true;
+
+              server.emit('open');
+
+              break;
+
+            default:
+              return false;
+          }
+
+          server.isOpening = false;
+
+          if (err === null) {
+            resolve(null);
+          }
+          else {
+            reject(err);
+          }
+
+          return true;
+        };
+
+        const dataHandler = Mongod.getTextLineAggregator((value) => {
+          const matches = value.match(keyRE);
+
+          if (matches !== null) {
+            for (let match of matches) {
+              if (matchHandler(match, value)) {
+                server.process.stdout.removeListener('data', dataHandler);
+                server.process.stderr.removeListener('data', dataHandler);
+
+                return;
+              }
             }
           }
-        }
+        });
+
+        const exitHandler = () => {
+          server.close();
+        };
+
+        const getDataPropagator = (event) =>
+          Mongod.getTextLineAggregator((line) => server.emit(event, line));
+
+        server.emit('opening');
+
+        server.process = childprocess.spawn(
+          server.config.bin,
+          Mongod.parseFlags(server.config)
+        );
+
+        server.process.stderr.on('data', dataHandler);
+        server.process.stderr.on('data', getDataPropagator('stderr'));
+        server.process.stdout.on('data', dataHandler);
+        server.process.stdout.on('data', getDataPropagator('stdout'));
+        server.process.on('close', () => {
+          server.process = null;
+          server.port = null;
+          server.pid = null;
+          server.isRunning = false;
+          server.isClosing = false;
+
+          process.removeListener('exit', exitHandler);
+          server.emit('close');
+        });
+        process.on('exit', exitHandler);
       });
-
-      const exitHandler = () => {
-        server.close();
-      };
-
-      const getDataPropagator = (event) =>
-        Mongod.getTextLineAggregator((line) => server.emit(event, line));
-
-      server.isOpening = true;
-
-      server.emit('opening');
-
-      server.process = childprocess.spawn(
-        server.config.bin,
-        Mongod.parseFlags(server.config)
-      );
-
-      server.process.stderr.on('data', dataHandler);
-      server.process.stderr.on('data', getDataPropagator('stderr'));
-      server.process.stdout.on('data', dataHandler);
-      server.process.stdout.on('data', getDataPropagator('stdout'));
-      server.process.on('close', () => {
-        server.process = null;
-        server.port = null;
-        server.pid = null;
-        server.isRunning = false;
-        server.isClosing = false;
-
-        process.removeListener('exit', exitHandler);
-        server.emit('close');
-      });
-      process.on('exit', exitHandler);
     });
 
     return server.openPromise;
@@ -263,34 +257,27 @@ class Mongod extends events.EventEmitter {
    * Stop a given {@link Mongod}.
    * @protected
    * @argument {Mongod} server
-   * @argument {Boolean} isCallback
    * @return {Promise}
    */
-  static close(server, isCallback) {
-    if (server.isOpening) {
-      server.closePromise = new Promise((resolve, reject) => {
-        const close = () =>
-          exports.close(server, true).then(resolve).catch(reject);
+  static close(server) {
+    if (server.isClosing) {
+      return server.closePromise;
+    }
 
-        server.isOpening = false;
-        server.isClosing = true;
+    server.isClosing = true;
+    server.isOpening = false;
+    server.closePromise = server.promiseQueue.add(() => {
+      if (server.isOpening || !server.isRunning) {
+        server.isClosing = false;
 
-        server.openPromise.then(close).catch(close);
+        return Promise.resolve(null);
+      }
+
+      return new Promise((resolve) => {
+        server.emit('closing');
+        server.process.once('close', () => resolve(null));
+        server.process.kill();
       });
-
-      return server.closePromise;
-    }
-
-    if (server.isClosing && !isCallback || !server.isRunning) {
-      return server.closePromise;
-    }
-
-    server.closePromise = new Promise((resolve) => {
-      server.isClosing = true;
-
-      server.emit('closing');
-      server.process.once('close', () => resolve(null));
-      server.process.kill();
     });
 
     return server.closePromise;
@@ -350,6 +337,13 @@ class Mongod extends events.EventEmitter {
      * @type {Promise}
      */
     this.closePromise = Promise.resolve(null);
+
+    /**
+     * A serial queue of open and close promises.
+     * @protected
+     * @type {PromiseQueue}
+     */
+    this.promiseQueue = new PromiseQueue(1);
 
     /**
      * Determine if the instance is closing a MongoDB server; {@linkcode true}
